improve(plugin): expose logger for user lib
